Rename ActionIcon and Header component, styles and stories
