+++ react-app/src/components/ActionIcon/stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 
-import ActionIcon from './ActionIcon';
+import ActionIcon from './';
 
 storiesOf('ActionIcon', module)
   .add('ActionIcon back', () => (
